move tracking consent logic into TrackingConsentRequester
